requirements-dev.txt
ansible>=9
ansible-lint
molecule
molecule-plugins[docker]
yamllint
pytest
markdownlint-cli
PyYAML
jinja2